refactor [James]: refactor merchandise route
